+++ changelog.md
@@ -42,6 +42,7 @@
 * [1746](https://github.com/zeta-chain/node/pull/1746) - rename smoke tests to e2e tests
 * [1753](https://github.com/zeta-chain/node/pull/1753) - fix gosec errors on usage of rand package
 * [1762](https://github.com/zeta-chain/node/pull/1762) - improve coverage for fungibile module
+* [1782](https://github.com/zeta-chain/node/pull/1782) - improve coverage for fungibile module system contract
 
 ### CI